requirements.txt
Flask
openai-whisper
translatepy
edge-tts
pysrt
ffmpeg-python
